% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCEAweb.R
\name{launch}
\alias{launch}
\title{Internal launch function}
\usage{
launch(e, c, parameters, ...)
}
\arguments{
\item{e}{A matrix containing the simulations for the effectiveness
variable (with number of simulation rows and number of interventions
columns). Defaults at NULL, which means the user has to load their
own values using the web-interface}

\item{c}{A matrix containing the simulations for the cost
variable (with number of simulation rows and number of interventions
columns). Defaults at NULL, which means the user has to load their
own values using the web-interface}

\item{parameters}{A matrix with the simulations for all the relevant
model parameters. Defaults at NULL, which means the user has to load
their own values using the web-interface}

\item{...}{Additional parameters.}
}
\value{
runApp
}
\description{
Internal launch function
}
