% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{Smoking_parameters}
\alias{Smoking_parameters}
\title{Smoking Cessation Example - Parameters}
\format{
A CSV file with 500 rows and 4 variables:
\describe{
\item{p1}{Simulations for the probability of quitting in intervention 1}
\item{p2}{Simulations for the probability of quitting in intervention 2}
\item{p3}{Simulations for the probability of quitting in intervention 3}
\item{p4}{Simulations for the probability of quitting in intervention 4}
}
}
\source{
Generated for package examples
}
\description{
This CSV file contains simulated parameter data for the Smoking Cessation example.
}
\examples{
\dontrun{
csv_file <- system.file("extdata", "smoking_parameters.csv", package = "BCEAweb")
data <- read.csv(csv_file)
head(data)
}
}
