% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{Smoking_results}
\alias{Smoking_results}
\title{Smoking Cessation Example - Results}
\format{
A CSV file with 500 rows and 4 variables:
\describe{
\item{e.No.treatment}{Simulations for the effects in intervention 1}
\item{c.No.treatment}{Simulations for the costs in intervention 1}
\item{e.Self.help}{Simulations for the effects in intervention 2}
\item{c.Self.help}{Simulations for the costs in intervention 2}
\item{e.Individual.counselling}{Simulations for the effects in intervention 3}
\item{c.Individual.counselling}{Simulations for the effects in intervention 3}
\item{e.Group.counselling}{Simulations for the effects in intervention 4}
\item{c.Group.counselling}{Simulations for the costs in intervention 4}
}
}
\source{
Generated for package examples
}
\description{
This CSV file contains simulated outputs for the Smoking Cessation example.
}
\examples{
\dontrun{
csv_file <- system.file("extdata", "Smoking_results.csv", package = "BCEAweb")
data <- read.csv(csv_file)
head(data)
}
}
