% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{Vaccine_results}
\alias{Vaccine_results}
\title{Vaccine Example - Results}
\format{
A CSV file with 500 rows and 2 variables:
\describe{
\item{e.Status.Quo}{Simulations for the effects in intervention 1}
\item{c.Status.Quo}{Simulations for the costs in intervention 1}
\item{e.Vaccination}{Simulations for the effects in intervention 2}
\item{c.Vaccination}{Simulations for the costs in intervention 2}
}
}
\source{
Generated for package examples
}
\description{
This CSV file contains simulated outputs for the Vaccine example.
}
\examples{
\dontrun{
csv_file <- system.file("extdata", "Vaccine_results.csv", package = "BCEAweb")
data <- read.csv(csv_file)
head(data)
}
}
